[chore] 컨트롤러 명 변경 및 correctAnser float형으로 반환하도록 변경

+++ src/main/java/com/tmax/ai/dto/response/TestResponseDto.java
@@ -12,6 +12,6 @@ public class TestResponseDto {
 
     private Boolean isCorrect;
     @JsonInclude(JsonInclude.Include.NON_NULL) // null인 경우 json 반환시 제거
-    private Integer correctAnswer;
+    private Float correctAnswer;
 
 }
